django/djangoCeleryRabbitMQRedisApp/requirements.txt
django==1.10.4
celery==3.1.20  
djangorestframework==3.3.1  
psycopg2==2.5.3  
redis==2.10.5  
